Issue #1: Add PROXY protocol support to tcp+sni
The TCP+SNI proxy fails when used behind an AWS gateway
with PROXY protocol enabled since it will not properly
decode the ClientHello from the TLS handshake.

Adding the proxyproto listener support to the TCP listener
should fix this as it should properly strip the PROXY header
from the bytestream.